+++ pypef/settings.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+# Created on 05 October 2020
+# @authors: Niklas Siedhoff, Alexander-Maurice Illig
+# @contact: <niklas.siedhoff@rwth-aachen.de>
+# PyPEF - Pythonic Protein Engineering Framework
+# https://github.com/niklases/PyPEF
+# Licensed under Creative Commons Attribution-ShareAlike 4.0 International Public License (CC BY-SA 4.0)
+# For more information about the license see https://creativecommons.org/licenses/by-nc/4.0/legalcode
+
+# PyPEF – An Integrated Framework for Data-Driven Protein Engineering
+# Journal of Chemical Information and Modeling, 2021, 61, 3463-3476
+# https://doi.org/10.1021/acs.jcim.1c00099
+
+
+USE_RAY = False
